+++ android/app/src/main/java/com/zzang/chongdae/common/firebase/fcm/NotificationImportance.kt
@@ -1,17 +1,21 @@
 package com.zzang.chongdae.common.firebase.fcm
 
-//enum class NotificationImportance {
-//    DEFAULT,
-//    HIGH,
-//    ;
-//
-//    companion object {
-//        fun of(importance: Int): NotificationImportance {
-//            return when(importance) {
-//                3 -> DEFAULT
-//                4 -> HIGH
-//                else -> error("notification importance가 유효하지 않습니다.")
-//            }
-//        }
-//    }
-//}
+import android.app.NotificationManager
+
+sealed interface NotificationImportance {
+    val importance: Int
+    val channelId: String
+    val channelName: String
+
+    data class Default(
+        override val importance: Int = NotificationManager.IMPORTANCE_DEFAULT,
+        override val channelId: String = "channel_importance_default",
+        override val channelName: String = "Default Channel",
+    ) : NotificationImportance
+
+    data class High(
+        override val importance: Int = NotificationManager.IMPORTANCE_HIGH,
+        override val channelId: String = "channel_importance_high",
+        override val channelName: String = "High Channel",
+    ) : NotificationImportance
+}
